+++ src/config/extends/eslint-plugin-jsdoc.d.ts
