+++ src/config/extends/eslint-plugin-mdx.d.ts
