+++ src/config/extends/eslint-plugin-n.d.ts
