+++ src/config/extends/eslint-plugin-prettier.d.ts
