+++ src/config/extends/eslint-plugin-vue.d.ts
