+++ src/config/extends/intlify-vue-i18n.d.ts
