+++ src/config/extends/typescript-eslint.d.ts
