+++ src/config/index.d.ts
