+++ src/config/plugin.d.ts
