+++ src/config/settings/mdx.d.ts
