+++ src/flat-config/index.d.ts
@@ -4,11 +4,11 @@ import type { LanguageOptions } from './language-options';
 import type { LinterOptions } from './linter-options';
 
 /**
- * ESLint Flat Configuration.
+ * Flat ESLint Configuration.
  *
  * @see [Configuration Files (New)](https://eslint.org/docs/latest/user-guide/configuring/configuration-files-new)
  */
-export interface ESLintFlatConfig {
+export interface FlatESLintConfig {
   /**
    * An array of glob patterns indicating the files that the configuration object should apply to. If not specified, the configuration object applies to all files.
    *
