+++ src/flat-config/linter-options.d.ts
